[NUI] Change Parameter type of KeyboardRepeatInfo APIs

+++ src/Tizen.NUI/src/internal/ManualPINVOKE.cs
@@ -303,10 +303,10 @@ namespace Tizen.NUI
         public static extern bool UngrabKey(System.IntPtr Window, int DaliKey);
 
         [global::System.Runtime.InteropServices.DllImport("libdali-csharp-binder.so", EntryPoint = "CSharp_Dali_SetKeyboardRepeatInfo")]
-        public static extern bool SetKeyboardRepeatInfo(double rate, double delay);
+        public static extern bool SetKeyboardRepeatInfo(float rate, float delay);
 
         [global::System.Runtime.InteropServices.DllImport("libdali-csharp-binder.so", EntryPoint = "CSharp_Dali_GetKeyboardRepeatInfo")]
-        public static extern bool GetKeyboardRepeatInfo(out double rate, out double delay);
+        public static extern bool GetKeyboardRepeatInfo(out float rate, out float delay);
 
         [global::System.Runtime.InteropServices.DllImport("libdali-csharp-binder.so", EntryPoint = "CSharp_Dali_GetNativeWindowHandler")]
         public static extern System.IntPtr GetNativeWindowHandler(System.IntPtr Window);
